Added a page on org-info-js.
This is primarily a test of my first Worg contribution.

NOTE that I've put in an initial attempt on how you might use Worg to
manage bugs and feature requests for org-info-js.

Essentially there are two sequences, one for features the other for
bugs.  Bugs progress like so:

BUGREPORT -> BUG -> FIXED

Features like this:

SUGGESTION -> TOIMPLEMENT -> IMPLEMENTED

If it's a crazy idea, then feel free to ignore, delete, or edit it.

Also added Sebastian's demonstration of org-info-js as powerpoint.